price data type change to flaot from double or int

+++ src/main/java/sfa/product_service/entity/ProductPriceEntity.java
@@ -14,8 +14,8 @@ import lombok.experimental.FieldDefaults;
 @Table(name = "product_price_table")
 public class ProductPriceEntity extends BaseEntity{
     Long productId;
-    Double wareHousePrice;
-    Double stockListPrice;
-    Double retailerPrice;
-    Double gst;
+    float wareHousePrice;
+    float stockListPrice;
+    float retailerPrice;
+    float gst;
 }
